feat: Setup frontend and api in same repo
This allows a dev to build both a frontend and its corresponding backend
api in the same repo.  This setup is different than is usually used for
real projects (usually the API is split into a different repo than the
frontend due to deployment difference), but it is a very good setup for
demo or test projects.

Locally you can run the frontend and API simultaneously in 2 different
terminals using `npm run api` and `npm run frontend`.  This will allow
you to test the full project locally.

In addition, if needed you should be able to deploy this project to a
web hosting service (e.g. Heroku) with minimal to no changes.

General structure of this setup was provided by this article:

https://www.freecodecamp.org/news/how-to-create-a-react-app-with-a-node-backend-the-complete-guide/