feat: Include SQLite as built in API data store
This feature adds support for using SQLite as the "embedded" data store
for the API in this project.  Upon running the API for the first time,
a new database will be created if one does not already exist.  It will
create two tables in the database: inbound_leads and departments

The database is ignored as part of the .gitignore setup so it will not
be checked in.

The majority of the setup for this feature can be found here:

https://developerhowto.com/2018/12/29/build-a-rest-api-with-node-js-and-express-js/